app/requirements.txt
APScheduler==3.5.3
beautifulsoup4==4.7.1
Flask==1.0.2
requests==2.21.0
psycopg2-binary==2.7.7
lxml==4.3.1